Move illustration to root package
