+++ app/src/main/java/org/simple/clinic/AppDatabase.kt
@@ -7,8 +7,8 @@ import org.simple.clinic.bp.BloodPressureMeasurement
 import org.simple.clinic.drugs.PrescribedDrug
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.home.overdue.OverdueAppointment
-import org.simple.clinic.home.patients.illustration.DayOfMonth
-import org.simple.clinic.home.patients.illustration.HomescreenIllustration
+import org.simple.clinic.illustration.DayOfMonth
+import org.simple.clinic.illustration.HomescreenIllustration
 import org.simple.clinic.medicalhistory.Answer
 import org.simple.clinic.medicalhistory.MedicalHistory
 import org.simple.clinic.overdue.Appointment
