+++ app/src/main/java/org/simple/clinic/illustration/HomescreenIllustration.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.home.patients.illustration
+package org.simple.clinic.illustration
 
 import androidx.room.Dao
 import androidx.room.Embedded
