+++ app/src/test/java/org/simple/clinic/home/patients/PatientsScreenControllerTest.kt
@@ -23,7 +23,7 @@ import org.mockito.Mockito.verify
 import org.simple.clinic.activity.TheActivityLifecycle
 import org.simple.clinic.appupdate.AppUpdateState
 import org.simple.clinic.appupdate.CheckAppUpdateAvailability
-import org.simple.clinic.home.patients.illustration.HomescreenIllustrationRepository
+import org.simple.clinic.illustration.HomescreenIllustrationRepository
 import org.simple.clinic.patient.PatientConfig
 import org.simple.clinic.patient.PatientMocker
 import org.simple.clinic.storage.files.GetFileResult
