+++ app/src/test/java/org/simple/clinic/illustration/DayOfMonthMoshiTypeAdapterTest.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.home.patients.illustration
+package org.simple.clinic.illustration
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
